replace old training with Tim's ttraining
